+++ R/splitstackshape_helper_functions.R
@@ -66,7 +66,7 @@ all_names <- function(current_names, stubs, end_stub = FALSE, ids = NULL,
   }
   
   id_names <- ids
-
+  
   levs <- unique(gsub(paste(stubs, collapse = "|"), "", stub_names))
   stub_levs <- trim_vec(CJ(stubs, levs)[, if (end_stub) paste0(V2, V1) else paste0(V1, V2)])
   full_names <- c(id_names, stub_levs[order(stub_levs)])